made diffrent file for gp queries
