+++ services/index.js
@@ -1,38 +1,9 @@
-import { request, gql } from 'graphql-request';
-
+import { request, gql } from "graphql-request";
+import * as graphQL from "../Graphql/index";
 const graphqlAPI = process.env.NEXT_PUBLIC_GRAPHCMS_ENDPOINT;
 
 export const getPosts = async () => {
-  const query = gql`
-    query MyQuery {
-      postsConnection {
-        edges {
-          cursor
-          node {
-            author {
-              bio
-              name
-              id
-              photo {
-                url
-              }
-            }
-            createdAt
-            slug
-            title
-            excerpt
-            featuredImage {
-              url
-            }
-            categories {
-              name
-              slug
-            }
-          }
-        }
-      }
-    }
-  `;
+  const query = graphQL.posts;
 
   const result = await request(graphqlAPI, query);
 
@@ -40,14 +11,7 @@ export const getPosts = async () => {
 };
 
 export const getCategories = async () => {
-  const query = gql`
-    query GetGategories {
-        categories {
-          name
-          slug
-        }
-    }
-  `;
+  const query = graphQL.categories;
 
   const result = await request(graphqlAPI, query);
 
@@ -55,33 +19,7 @@ export const getCategories = async () => {
 };
 
 export const getPostDetails = async (slug) => {
-  const query = gql`
-    query GetPostDetails($slug : String!) {
-      post(where: {slug: $slug}) {
-        title
-        excerpt
-        featuredImage {
-          url
-        }
-        author{
-          name
-          bio
-          photo {
-            url
-          }
-        }
-        createdAt
-        slug
-        content {
-          raw
-        }
-        categories {
-          name
-          slug
-        }
-      }
-    }
-  `;
+  const query = graphQL.details;
 
   const result = await request(graphqlAPI, query, { slug });
 
@@ -89,92 +27,21 @@ export const getPostDetails = async (slug) => {
 };
 
 export const getSimilarPosts = async (categories, slug) => {
-  const query = gql`
-    query GetPostDetails($slug: String!, $categories: [String!]) {
-      posts(
-        where: {slug_not: $slug, AND: {categories_some: {slug_in: $categories}}}
-        last: 3
-      ) {
-        title
-        featuredImage {
-          url
-        }
-        createdAt
-        slug
-      }
-    }
-  `;
+  const query = graphQL.similarPosts;
   const result = await request(graphqlAPI, query, { slug, categories });
 
   return result.posts;
 };
 
 export const getAdjacentPosts = async (createdAt, slug) => {
-  const query = gql`
-    query GetAdjacentPosts($createdAt: DateTime!,$slug:String!) {
-      next:posts(
-        first: 1
-        orderBy: createdAt_ASC
-        where: {slug_not: $slug, AND: {createdAt_gte: $createdAt}}
-      ) {
-        title
-        featuredImage {
-          url
-        }
-        createdAt
-        slug
-      }
-      previous:posts(
-        first: 1
-        orderBy: createdAt_DESC
-        where: {slug_not: $slug, AND: {createdAt_lte: $createdAt}}
-      ) {
-        title
-        featuredImage {
-          url
-        }
-        createdAt
-        slug
-      }
-    }
-  `;
-
+  const query = graphQL.adjacentPosts;
   const result = await request(graphqlAPI, query, { slug, createdAt });
 
   return { next: result.next[0], previous: result.previous[0] };
 };
 
 export const getCategoryPost = async (slug) => {
-  const query = gql`
-    query GetCategoryPost($slug: String!) {
-      postsConnection(where: {categories_some: {slug: $slug}}) {
-        edges {
-          cursor
-          node {
-            author {
-              bio
-              name
-              id
-              photo {
-                url
-              }
-            }
-            createdAt
-            slug
-            title
-            excerpt
-            featuredImage {
-              url
-            }
-            categories {
-              name
-              slug
-            }
-          }
-        }
-      }
-    }
-  `;
+  const query = categoryPost;
 
   const result = await request(graphqlAPI, query, { slug });
 
@@ -182,24 +49,7 @@ export const getCategoryPost = async (slug) => {
 };
 
 export const getFeaturedPosts = async () => {
-  const query = gql`
-    query GetCategoryPost() {
-      posts(where: {featuredPost: true}) {
-        author {
-          name
-          photo {
-            url
-          }
-        }
-        featuredImage {
-          url
-        }
-        title
-        slug
-        createdAt
-      }
-    }   
-  `;
+  const query = graphQL.featuredPosts;
 
   const result = await request(graphqlAPI, query);
 
@@ -207,10 +57,10 @@ export const getFeaturedPosts = async () => {
 };
 
 export const submitComment = async (obj) => {
-  const result = await fetch('/api/comments', {
-    method: 'POST',
+  const result = await fetch("/api/comments", {
+    method: "POST",
     headers: {
-      'Content-Type': 'application/json',
+      "Content-Type": "application/json",
     },
     body: JSON.stringify(obj),
   });
@@ -220,8 +70,8 @@ export const submitComment = async (obj) => {
 
 export const getComments = async (slug) => {
   const query = gql`
-    query GetComments($slug:String!) {
-      comments(where: {post: {slug:$slug}}){
+    query GetComments($slug: String!) {
+      comments(where: { post: { slug: $slug } }) {
         name
         createdAt
         comment
